Import 2.3.11pre6

+++ fs/nfs/read.c
@@ -152,7 +152,7 @@ nfs_readpage_result(struct rpc_task *task)
 		fail++;
 		dprintk("NFS: %d successful reads, %d failures\n", succ, fail);
 	}
-	page->owner = (int)current; // HACK, FIXME, will go away.
+	page->owner = current; // HACK, FIXME, will go away.
 	UnlockPage(page);
 	free_page(address);
